decoder: add assert to verify that time hasn't gone backwards
Events are logged with clock time, which can go backwards when the
clock is reset or resynched.  Our analysis code assumes that the
timestamp ordering is consistent with event occurance ordering, so we
need to reject traces with bad timestamps.

This commit adds an assert to verify that time has not gone backwards.
Should we find traces that do go backwards, we can then figure out how
to deal with them.