install libget_time.so to system/lib
JNI shared libraries are by default placed into the lib folder of the
apk and then extracted into the /data partition upon installation.
System apps are never 'installed', so the extraction does not take
place.  Instead, the JNI shared libraries must be placed into
/system/lib/ at build time. This commit ensures the JNI shared library
is installed by specifying it as a required module.